+++ src/App.module.css
@@ -3,7 +3,6 @@
     height: 100%;
     display: flex;
     flex-direction: column;
-    align-items: center;
-    justify-content: center;
+  
     font-size: 1rem;
 }
